test#24: 파일명 변경 및 코드 수정
